+++ .github/workflows/build.yml
@@ -1,15 +1,10 @@
 name: Build Crate
+run-name: ${{ inputs.crate }}
 
 on:
   workflow_call:
     inputs:
       crate:
-        type: string
-        required: true
-      license:
-        type: string
-        required: true
-      bins:
         type: string
         required: true
 
@@ -83,7 +78,7 @@ jobs:
         working-directory: ./build
         run: cargo build --release
       - name: Collect
-        run: ./collect.py ${{ inputs.bins }} ${{ inputs.license }} x86_64-unknown-linux-gnu ./build ./build/target/release
+        run: ./collect.py x86_64-unknown-linux-gnu ./build ./build/target/release
       - name: Artifact
         uses: actions/upload-artifact@v3
         with:
